Expose the internal guts of the haskell library
This can be useful for downstream consumers of the library who are OK
with depending on an unsafe and unstable interface.
