+++ nix-thunk.cabal
@@ -1,6 +1,6 @@
 cabal-version:      >=1.10
 name:               nix-thunk
-version:            0.6.0.0
+version:            0.6.1.0
 license:            BSD3
 license-file:       LICENSE
 copyright:          Obsidian Systems LLC 2020-2022
@@ -24,6 +24,7 @@ tested-with: GHC ==8.6.5 || ==8.8.4
 library
   exposed-modules:
     Nix.Thunk
+    Nix.Thunk.Internal
     Nix.Thunk.Command
 
   hs-source-dirs:   src
